Renamed utils to src
