+++ src/merkle.js
